Move verify_series_and_expr_api into test_series to prevent import struggles + add kwargs
